requirements.txt
Flask==2.3.3
Flask-Cors==3.0.10
torch==2.0.1
torchvision==0.15.2
Pillow==10.0.1
numpy==1.26.0
google-generativeai==0.2.0